// src/app/shared/interfaces/register-client-person.interface.ts
export interface RegisterClientPerson {
    email:            string;
    password:         string;
    username:         string;
    direccionEmpresa: string;
    apellidos:        string;
    fechaNacimiento:  string;
    nombres:          string;
    telefono:         string;
    numeroDocumento:  string;
}